# bin/gestao-hospitalar-swagger.yml
openapi: 3.0.1
info:
  title: Gestão Hospitalar API
  description: API para gestão de hospital
  contact:
    name: Yuri Nascimento
    url: https://github.com/yuriidiiego
    email: yuriidiiego@gmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: "1.0"
servers:
  - url: http://localhost:8080
    description: Local server
security:
  - bearerAuth: []
...
paths:
  /pacientes/{id}:
    get:
      tags:
        - Paciente
      summary: Busca paciente por id
      operationId: buscarPorId
      parameters:
        - name: id
          in: path
          description: Id do paciente
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      responses:
        "200":
          description: Paciente encontrado com sucesso
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PacienteResponse"
        "404":
          description: Paciente não encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      tags:
        - Paciente
      summary: Atualiza um paciente
      operationId: atualizar
      parameters:
        - name: id
          in: path
          description: Id do paciente
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PacientePutRequest"
        required: true
      responses:
        "200":
          description: Paciente atualizado com sucesso
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PacienteResponse"
        "400":
          description: Erro de validação
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Paciente não encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      tags:
        - Paciente
      summary: Deleta um paciente
      operationId: deletar
      parameters:
        - name: id
          in: path
          description: Id do paciente
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      responses:
        "204":
          description: Paciente deletado com sucesso
        "404":
          description: Paciente não encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      tags:
        - Paciente
      summary: Atualiza parcialmente um paciente
      operationId: atualizarParcial
      parameters:
        - name: id
          in: path
          description: Id do paciente
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PacientePatchRequest"
        required: true
      responses:
        "200":
          description: Paciente atualizado com sucesso
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PacienteResponse"
        "400":
          description: Erro de validação
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Paciente não encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /medicos/{id}:
    get:
      tags:
        - Médico
      summary: Busca médico por ID
      description: Retorna um médico pelo ID
      operationId: buscaMedicoPorId
      parameters:
        - name: id
          in: path
          description: ID do médico
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      responses:
        "200":
          description: Médico encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MedicoResponse"
        "404":
          description: Médico não encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      tags:
        - Médico
      summary: Atualiza médico
      operationId: atualizar_1
      parameters:
        - name: id
          in: path
          description: ID do médico
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MedicoPutRequest"
        required: true
      responses:
        "500":
          description: Erro interno
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "200":
          description: Médico atualizado com sucesso
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MedicoResponse"
        "404":
          description: Médico não encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      tags:
        - Médico
      summary: Deleta um médico
      operationId: deletar_1
      parameters:
        - name: id
          in: path
          description: ID do médico
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      responses:
        "500":
          description: Erro interno
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "204":
          description: Médico deletado com sucesso
        "404":
          description: Médico não encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      tags:
        - Médico
      summary: Atualiza parcialmente um médico
      operationId: atualizarParcial_1
      parameters:
        - name: id
          in: path
          description: ID do médico
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MedicoPatchRequest"
        required: true
      responses:
        "500":
          description: Erro interno
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "200":
          description: Médico atualizado com sucesso
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MedicoResponse"
        "404":
          description: Médico não encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /pacientes:
    get:
      tags:
        - Paciente
      summary: Busca todos os pacientes
      operationId: buscarTodos
      responses:
        "200":
          description: Pacientes encontrados com sucesso
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PacienteResponse"
    post:
      tags:
        - Paciente
      summary: Cadastra um paciente
      operationId: salvar
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PacientePostRequest"
        required: true
      responses:
        "400":
          description: Erro de validação
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "201":
          description: Paciente cadastrado com sucesso
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PacienteResponse"
...
  /pacientes/pdf:
    get:
      tags:
        - Paciente
      summary: gera um pdf com todos os pacientes
      operationId: gerarPdfPacientes
      responses:
        "500":
          description: erro interno no servidor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "200":
          description: pdf gerado com sucesso
          content:
            application/pdf: {}
  /pacientes/medicos/{idMedico}:
    get:
      tags:
        - Paciente
      summary: " Lista os pacientes de um determinado médico"
      operationId: buscarPorMedico
      parameters:
        - name: idMedico
          in: path
          description: Id do médico
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      responses:
        "200":
          description: Pacientes encontrados com sucesso
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PacienteResponse"
        "404":
          description: Médico não encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /medicos/pacientes/{idPaciente}:
    get:
      tags:
        - Médico
      summary: Lista médicos que atenderam um determinado paciente
      operationId: buscarPorPaciente
      parameters:
        - name: idPaciente
          in: path
          description: ID do paciente
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      responses:
        "500":
          description: Erro interno
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "200":
          description: Médicos encontrados com sucesso
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MedicoResponse"
        "404":
          description: Paciente não encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
    PacientePutRequest:
      type: object
      properties:
        nome:
          type: string
        cpf:
          type: string
        dataNascimento:
          type: string
          format: date
        sexo:
          type: string
          enum:
            - MASCULINO
            - FEMININO
    PacienteResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nome:
          type: string
        cpf:
          type: string
        dataNascimento:
          type: string
          format: date
        sexo:
          type: string
    ErrorResponse:
      type: object
      properties:
        statuscode:
          type: integer
          format: int32
        timestamp:
          type: string
          format: date-time
        erro:
          type: string
    MedicoPutRequest:
      type: object
      properties:
        nome:
          type: string
        cpf:
          type: string
        dataNascimento:
          type: string
          format: date
        crm:
          type: string
        sexo:
          type: string
          enum:
            - MASCULINO
            - FEMININO
    MedicoResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nome:
          type: string
        cpf:
          type: string
        crm:
          type: string
        dataNascimento:
          type: string
          format: date
        sexo:
          type: string
          enum:
            - MASCULINO
            - FEMININO
...
    PacientePostRequest:
      required:
        - cpf
        - dataNascimento
        - nome
        - sexo
      type: object
      properties:
        nome:
          type: string
          description: Nome do paciente
          example: João da Silva
        cpf:
          type: string
          description: CPF do paciente
          example: "70989098230"
        dataNascimento:
          pattern: dd/MM/yyyy
          type: string
          description: Data de nascimento do paciente
          example: 02/02/1980
        sexo:
          type: string
          description: Sexo do paciente
          example: MASCULINO
          enum:
            - MASCULINO
            - FEMININO
...
    PacientePatchRequest:
      required:
        - cpf
        - dataNascimento
        - nome
        - sexo
      type: object
      properties:
        nome:
          type: string
          description: Nome do paciente
          example: João da Silva
        cpf:
          type: string
          description: CPF do paciente
          example: "70989098230"
        dataNascimento:
          pattern: dd/MM/yyyy
          type: string
          description: Data de nascimento do paciente
          example: 02/02/1980
        sexo:
          type: string
          description: Sexo do paciente
          example: MASCULINO
          enum:
            - MASCULINO
            - FEMININO
    MedicoPatchRequest:
      type: object
      properties:
        nome:
          type: string
        cpf:
          type: string
        dataNascimento:
          type: string
          format: date
        crm:
          type: string
        sexo:
          type: string
          enum:
            - MASCULINO
            - FEMININO
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT